requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
matplotlib
seaborn
joblib
scikit-surprise
fuzzywuzzy
requests